#39
Adds new methods to store an input room by getting that room from an Energy Grid. The previous method only allowed a room to be stored from a house, and was insufficient to meet the demands of US149, which requires the RoomList of House and Grid to be handled separately.