Fixes #267.
Coverage on EnergySettingsController is now 100%.

UpdateGrid method deleted - .save() method called in addGridToHouse
already updates the existing grid if present.

Changed addGridToHouse method to return the grid that was added.

Changed addGridToHouse method to not return null when the grid isn't
added for some reason, but instead throw a RuntimeException.

Changed all methods that call for addGridToHouse to catch the potential
RunTimeException.