Refs #286.
Adds tests for getRoomLogsInInterval unhappy path.

Fixes roomService updateHouseRoom method to not return an error convention, and instead throw a RunTimeException in its failure case. Fixes all methods that use updateHouseRoom to handle that RunTimeException, up until the UIs of the user stories that implement that method; those UIs now have a specific message for when that failure case is reached.