Fixes #453
Changes the path of all loggers generated in testing environment to use
a dump HTML file for all the things it logs. This allows us to not pollute
the actual HTML files every time we run the test suite. The dump file
can be used for debugging tests; the actual logging files can be used
for runtime.